Config fixes Dependencies cleanup and Update { - CitizensAPI - XConomy API - Vault API - Towny API }

+++ src/main/java/economy/pcconomy/backend/economy/bank/Bank.java
@@ -124,7 +124,7 @@ public class Bank extends Capitalist {
         var bigInflation = new AtomicReference<>(0d);
 
         GorodkiUniverse.getInstance().getNPCGorods().parallelStream().forEach((npc) -> {
-            bigInflation.updateAndGet(v -> (v + npc.getLocalInflation()));
+            bigInflation.updateAndGet(v -> (Double) (v + npc.getLocalInflation()));
             count.addAndGet(1);
         });
